refactor: rename component according to design specs
