--- src/components/DimensionPanel/index.js
@@ -1,1 +1,0 @@
-export { DimensionPanel } from './DimensionPanel.js'
