+++ src/components/MainSidebar/InputPanel/InputOption.js
