+++ src/components/MainSidebar/InputPanel/InputPanel.js
