+++ src/components/MainSidebar/InputPanel/InputPanel.module.css
