+++ src/components/MainSidebar/InputPanel/index.js
