+++ src/components/MainSidebar/MenuItem/index.js
@@ -1,0 +1,1 @@
+export { MenuItem } from './MenuItem.js'
