+++ src/components/MainSidebar/index.js
@@ -1,0 +1,1 @@
+export { MainSidebar } from './MainSidebar.js'
